Improves admin and product functionalities.
Adds authentication for admin login using correo and contraseña.

Implements product filtering by location (user and store), name, and stock.

Changes ranking ID type from Long to int.

+++ Grupo_11/src/main/java/com/DBD/Grupo_11/Controller/RankingController.java
@@ -17,7 +17,7 @@ public class RankingController {
     public Object saveRanking(@RequestBody Ranking ranking) {return rankingService.save(ranking);}
 
     @GetMapping("/findById/{id}")
-    public Object findById(@PathVariable Long id){return rankingService.getById(id);}
+    public Object findById(@PathVariable int id){return rankingService.getById(id);}
 
     @GetMapping("/findAll")
     public List<Ranking> findAll(){return rankingService.findAll();}
